size_t lines, quellage
svn:revision:39960
svn:branch:trunk
svn:account:brlcad

+++ src/conv/asc/asc2g.c
@@ -1063,7 +1063,7 @@ polyhbld(void)
     char	*cp;
     char	*name;
     long	startpos;
-    long	nlines;
+    size_t	nlines;
     struct rt_pg_internal	*pg;
     struct rt_db_internal	intern;
     struct bn_tol	tol;
